requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pandas
pyarrow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-pptx
xlsxwriter
python-dotenv
